+++ app/src/features/snippet/lib/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./canvasToBlob";
 export * from "./createCodeBlockEditor";
 export * from "./createCanvasElement";
+export * from "./insertTabSpace";
